Modularize L10N JSON File - Settings
