+++ gui/src/components/settings/ntp-setting.vue
@@ -39,12 +39,15 @@ import { Component, Vue, Prop } from "vue-property-decorator";
 import CortxDateTimeConfig from "../onboarding/system-config/date-time/date-time-config.vue";
 import { Api } from "../../services/api";
 import apiRegister from "../../services/api-register";
-import i18n from "../../i18n";
+import i18n from "./settings.json";
 
 @Component({
   name: "cortx-date-time",
   components: {
     cortxDateTimeConfig: CortxDateTimeConfig
+  },
+  i18n: {
+    messages: i18n
   }
 })
 export default class CortxNtpSetting extends Vue {
